refactor: split DashboardScreen.kt into focused per-card files
Purely mechanical extraction, zero behavior change. Code moved verbatim
into new files in the same package:

- CarImagePager.kt: CarSelectorPager, CarImage, glow bitmap loading
- BatteryCard.kt: BatteryCard, ChargingProgressBar,
  ChargingPowerGaugeCompact, ChargingDetailsRow, formatHoursMinutes,
  BatteryCard previews
- LocationCard.kt: LocationCard, LocationChip
- ActivityCard.kt: VehicleInfoCard, TripsHeroTile, NavButton
- TirePressureCard.kt: TirePressureCard, TyreSegment
- WhereWasIDateTimePicker.kt: WhereWasIDateTimePicker
- StatusIndicatorsRow.kt: StatusIcon, StatusIndicatorsRow,
  timestamp formatter helpers

Symbols referenced across the new files were widened from private to
internal (CarSelectorPager, BatteryCard, StatusIndicatorsRow,
LocationCard, VehicleInfoCard, TirePressureCard, WhereWasIDateTimePicker);
single-file helpers stay private.
